requirements.txt
flask
playwright
google-generativeai
python-dotenv
beautifulsoup4
